Update and rename test1 to win_lose
This is the first draft of Danger Zone that is being put on GitHub. The game displays the position values of each player throughout the game, and displays a winner. However, it does not limit the number of moves a player can make depending on their position, which is a huge issue. Edits from now on will be updated to GitHub.