Update and rename win_lose to all_rules
This new version, with a few more if statements added that limit the amount of spaces a player can move when they are close to the opposing player, allows the game to be played without loopholes. There are some bugs that need to be worked out still. For instance, the turn order goes p1, p2, and then p2, p1 for no reason. And, when going in reverse order, the moves are not registered. Sometimes turns are skipped also. Some more work still needs to be done.